third_party/WebKit/LayoutTests/animations/web-animations/neutral-keyframe-easing.html

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<div id="target"></div>
<script>
test(() => {
  target.style.left = '100px';
  var animation = target.animate([{easing: 'steps(2, start)'}, {left: '200px'}], 1000);
  animation.pause();
  animation.currentTime = 0;
  assert_equals(getComputedStyle(target).left, '150px');
}, 'Neutral keyframes should be able to specify their own keyframe timing function.');
</script>
